Moving function to utils
