+++ imap_processing/utils.py
@@ -606,3 +606,50 @@ def filter_day_boundary_data(dataset: xr.Dataset, start_date: str) -> xr.Dataset
     end_ttj2000ns = str_yyyymmdd_to_ttj2000ns(next_day) - 1
     logger.info(f"Filtering dataset out of day boundary of {start_date}.")
     return dataset.sel(epoch=slice(start_ttj2000ns, end_ttj2000ns))
+
+
+def check_epochs_within_day_offsets(
+    datasets: list[xr.Dataset],
+    day: np.datetime64,
+) -> None:
+    """
+    Raise an error if any dataset epoch falls more than 24 hours outside day.
+
+    A tolerance of ±24 hours around the expected processing day is allowed
+    to accommodate data that straddles midnight. Epochs beyond that window
+    may indicate the wrong input file was provided.  Eg.
+              day = "202605012"
+              lower = "20260511"
+              upper = "20260513"
+
+    If any data outside of this range is found, this function throws an error.
+    Some instruments can have buffer times beyond daily file date, but they should not
+    be more than 24hrs from the daily file date.
+
+    Parameters
+    ----------
+    datasets : list[xarray.Dataset]
+        Datasets whose ``epoch`` coordinate will be checked.
+    day : numpy.datetime64
+        The expected processing day.
+
+    Raises
+    ------
+    ValueError
+        If any epoch value is more than 24 hours before ``day`` or more
+        than 24 hours after the end of ``day``.
+    """
+    lower = str_yyyymmdd_to_ttj2000ns(
+        str(day - np.timedelta64(1, "D")).replace("-", "")
+    )
+    upper = str_yyyymmdd_to_ttj2000ns(
+        str(day + np.timedelta64(2, "D")).replace("-", "")
+    )
+    for dataset in datasets:
+        epoch_ns = dataset["epoch"].values
+        if np.any(epoch_ns < lower) or np.any(epoch_ns >= upper):
+            dataset_logical_id = dataset.attrs.get("Logical_source", "unknown dataset")
+            raise ValueError(
+                f"Data in {dataset_logical_id} contains epochs more than"
+                f" 24 hours outside the expected processing day {day}."
+            )
